+++ package.json
@@ -23,6 +23,6 @@
     "eslint-plugin-react-hooks": "^5.0.0",
     "eslint-plugin-react-refresh": "^0.4.13",
     "globals": "^15.11.0",
-    "vite": "^5.4.9"
+    "vite": "^5.4.10"
   }
 }
